+++ meridian/data/test_utils.py
@@ -409,10 +409,7 @@ DATASET_WITHOUT_TIME_VARIATION_IN_ORGANIC_MEDIA = xr.Dataset(
 )
 
 DATASET_WITHOUT_TIME_VARIATION_IN_ORGANIC_REACH = xr.Dataset(
-    coords=_REQUIRED_COORDS
-    | _MEDIA_COORDS
-    | _RF_COORDS
-    | _ORGANIC_RF_COORDS,
+    coords=_REQUIRED_COORDS | _MEDIA_COORDS | _RF_COORDS | _ORGANIC_RF_COORDS,
     data_vars=_REQUIRED_DATA_VARS
     | _MEDIA_DATA_VARS
     | _RF_DATA_VARS
@@ -818,11 +815,11 @@ def random_controls_da(
 
 def random_kpi_da(
     media: xr.DataArray,
-    controls: xr.DataArray,
     n_geos: int,
     n_times: int,
     n_media_channels: int,
-    n_controls: int,
+    n_controls: int | None = None,
+    controls: xr.DataArray | None = None,
     seed: int = 0,
     integer_geos: bool = False,
 ) -> xr.DataArray:
@@ -838,19 +835,26 @@ def random_kpi_da(
       n_media_channels,
       axis=2,
   )
-  control_geo_sd = abs(np.random.normal(0, 5, size=n_geos))
-  control_geo_sd = np.repeat(
-      np.repeat(control_geo_sd[:, np.newaxis], n_times, axis=1)[
-          ..., np.newaxis
-      ],
-      n_controls,
-      axis=2,
-  )
+  if n_controls:
+    control_geo_sd = abs(np.random.normal(0, 5, size=n_geos))
+    control_geo_sd = np.repeat(
+        np.repeat(control_geo_sd[:, np.newaxis], n_times, axis=1)[
+            ..., np.newaxis
+        ],
+        n_controls,
+        axis=2,
+    )
+  else:
+    control_geo_sd = 0
 
   # Simulates outcome which is the dependent variable. Typically this is the
   # number of units sold, but it can be any metric (e.g. revenue).
   media_portion = np.random.normal(media_common, media_geo_sd).sum(axis=2)
-  control_portion = np.random.normal(controls, control_geo_sd).sum(axis=2)
+  if controls is not None:
+    control_portion = np.random.normal(controls, control_geo_sd).sum(axis=2)
+  else:
+    control_portion = 0
+
   error = np.random.normal(0, 2, size=(n_geos, n_times))
   kpi = abs(media_portion + control_portion + error)
 
@@ -1161,7 +1165,7 @@ def random_dataset(
     n_geos: int,
     n_times: int,
     n_media_times: int,
-    n_controls: int,
+    n_controls: int | None = None,
     n_non_media_channels: int | None = None,
     n_organic_media_channels: int | None = None,
     n_organic_rf_channels: int | None = None,
@@ -1171,7 +1175,7 @@ def random_dataset(
     seed: int = 0,
     remove_media_time: bool = False,
     integer_geos: bool = False,
-):
+) -> xr.Dataset:
   """Generates a random dataset."""
   if n_media_channels:
     media = random_media_da(
@@ -1232,14 +1236,18 @@ def random_dataset(
   else:
     revenue_per_kpi = None
 
-  controls = random_controls_da(
-      media=media if n_media_channels else reach,
-      n_geos=n_geos,
-      n_times=n_times,
-      n_controls=n_controls,
-      seed=seed,
-      integer_geos=integer_geos,
-  )
+  if n_controls:
+    controls = random_controls_da(
+        media=media if n_media_channels else reach,
+        n_geos=n_geos,
+        n_times=n_times,
+        n_controls=n_controls,
+        seed=seed,
+        integer_geos=integer_geos,
+    )
+  else:
+    controls = None
+
   if n_non_media_channels:
     non_media_treatments = random_non_media_treatments_da(
         media=media if n_media_channels else reach,
@@ -1298,7 +1306,9 @@ def random_dataset(
       n_geos=n_geos, seed=seed, integer_geos=integer_geos
   )
 
-  dataset = xr.combine_by_coords([kpi, population, controls])
+  dataset = xr.combine_by_coords(
+      [kpi, population] + ([controls] if controls is not None else [])
+  )
   if revenue_per_kpi is not None:
     dataset = xr.combine_by_coords([dataset, revenue_per_kpi])
   if media is not None:
@@ -1347,7 +1357,7 @@ def random_dataset(
 
 def dataset_to_dataframe(
     dataset: xr.Dataset,
-    controls_column_names: list[str],
+    controls_column_names: list[str] | None = None,
     media_column_names: list[str] | None = None,
     media_spend_column_names: list[str] | None = None,
     reach_column_names: list[str] | None = None,
@@ -1396,10 +1406,15 @@ def dataset_to_dataframe(
   )
   population = dataset[c.POPULATION].to_dataframe(name=c.POPULATION)
 
-  controls = dataset[c.CONTROLS].to_dataframe(name=c.CONTROLS).unstack()
-  controls.columns = controls_column_names
+  if controls_column_names is not None:
+    controls = dataset[c.CONTROLS].to_dataframe(name=c.CONTROLS).unstack()
+    controls.columns = controls_column_names
+  else:
+    controls = None
 
-  result = kpi.join(revenue_per_kpi).join(population).join(controls)
+  result = kpi.join(revenue_per_kpi).join(population)
+  if controls is not None:
+    result = result.join(controls)
 
   if non_media_column_names is not None:
     non_media_treatments = (
@@ -1471,7 +1486,7 @@ def random_dataframe(
     n_geos,
     n_times,
     n_media_times,
-    n_controls,
+    n_controls=None,
     n_media_channels=None,
     n_rf_channels=None,
     seed=0,
@@ -1489,7 +1504,9 @@ def random_dataframe(
 
   return dataset_to_dataframe(
       dataset,
-      controls_column_names=_sample_names('control_', n_controls),
+      controls_column_names=(
+          _sample_names('control_', n_controls) if n_controls else None
+      ),
       media_column_names=_sample_names('media_', n_media_channels),
       media_spend_column_names=_sample_names('media_spend_', n_media_channels),
       reach_column_names=_sample_names('reach_', n_rf_channels),
@@ -1499,7 +1516,7 @@ def random_dataframe(
 
 
 def sample_coord_to_columns(
-    n_controls: int,
+    n_controls: int | None = None,
     n_media_channels: int | None = None,
     n_rf_channels: int | None = None,
     n_non_media_channels: int | None = None,
@@ -1550,7 +1567,7 @@ def sample_coord_to_columns(
       kpi=c.KPI,
       revenue_per_kpi=c.REVENUE_PER_KPI if include_revenue_per_kpi else None,
       population=c.POPULATION,
-      controls=_sample_names('control_', n_controls),
+      controls=(_sample_names('control_', n_controls) if n_controls else None),
       media=media,
       media_spend=media_spend,
       reach=reach,
@@ -1668,7 +1685,7 @@ def sample_input_data_non_revenue_revenue_per_kpi(
     n_geos: int = 10,
     n_times: int = 50,
     n_media_times: int = 53,
-    n_controls: int = 2,
+    n_controls: int | None = 2,
     n_non_media_channels: int | None = None,
     n_media_channels: int | None = None,
     n_rf_channels: int | None = None,
@@ -1694,10 +1711,10 @@ def sample_input_data_non_revenue_revenue_per_kpi(
       kpi_type=c.NON_REVENUE,
       revenue_per_kpi=dataset.revenue_per_kpi,
       population=dataset.population,
-      controls=dataset.controls,
-      non_media_treatments=dataset.non_media_treatments
-      if n_non_media_channels
-      else None,
+      controls=(dataset.controls if n_controls else None),
+      non_media_treatments=(
+          dataset.non_media_treatments if n_non_media_channels else None
+      ),
       media=dataset.media if n_media_channels else None,
       media_spend=dataset.media_spend if n_media_channels else None,
       reach=dataset.reach if n_rf_channels else None,
@@ -1705,9 +1722,9 @@ def sample_input_data_non_revenue_revenue_per_kpi(
       rf_spend=dataset.rf_spend if n_rf_channels else None,
       organic_media=dataset.organic_media if n_organic_media_channels else None,
       organic_reach=dataset.organic_reach if n_organic_rf_channels else None,
-      organic_frequency=dataset.organic_frequency
-      if n_organic_rf_channels
-      else None,
+      organic_frequency=(
+          dataset.organic_frequency if n_organic_rf_channels else None
+      ),
   )